v-menu pointer event tweaks
